Extract reusable workout summary table
