+++ app/src/components/layout/material-stack-header.tsx
@@ -9,6 +9,16 @@ export default function MaterialStackHeader(props: NativeStackHeaderProps) {
   const { isScrolled } = useScroll();
   const scrollColor = useRef(new Animated.Value(0)).current;
   const { colors } = useAppTheme();
+  const customHeaderLeft = props.options.headerLeft?.({
+    canGoBack: !!props.back,
+    tintColor: colors.onSurface,
+    label: props.back?.title,
+    href: undefined,
+  });
+  const showDefaultBackAction =
+    !!props.back &&
+    props.options.headerBackVisible !== false &&
+    customHeaderLeft === undefined;
 
   useEffect(() => {
     Animated.timing(scrollColor, {
@@ -26,10 +36,17 @@ export default function MaterialStackHeader(props: NativeStackHeaderProps) {
   return (
     <Animated.View style={{ backgroundColor }}>
       <Appbar.Header
-        mode={props.back && Platform.OS !== 'ios' ? 'small' : 'center-aligned'}
-        style={{ backgroundColor: 'transparent' }}
+        mode={
+          showDefaultBackAction && Platform.OS !== 'ios'
+            ? 'small'
+            : 'center-aligned'
+        }
+        style={{
+          backgroundColor: 'transparent',
+        }}
       >
-        {props.back ? (
+        {customHeaderLeft}
+        {showDefaultBackAction ? (
           <Appbar.BackAction onPress={() => props.navigation.goBack()} />
         ) : null}
         <Appbar.Content title={props.options.title} />
